+++ baker/DeployUtils.ts
@@ -9,9 +9,9 @@ import {
     writePendingFile,
     parseQueueContent,
 } from "./queue"
-import { DeployChange } from "./types"
 import { BAKED_SITE_DIR } from "settings/serverSettings"
 import { BAKED_BASE_URL } from "settings/clientSettings"
+import { DeployChange } from "clientUtils/owidTypes"
 
 const defaultCommitMessage = async (): Promise<string> => {
     let message = "Automated update"
