+++ clientUtils/owidTypes.ts
@@ -175,3 +175,15 @@ export enum DeployStatus {
     pending = "pending",
     // done = "done"
 }
+
+export interface DeployChange {
+    timeISOString?: string
+    authorName?: string
+    authorEmail?: string
+    message?: string
+}
+
+export interface Deploy {
+    status: DeployStatus
+    changes: DeployChange[]
+}
